main: Fix support for older polling accelerometers
Add support for per-axis accel scaling as required for older
accelerometers:
"
It used to be hard to have the same range out of the plane of the silicon
than within it, so was common to have sensors with different ranges and
hence scales in z direction from x and y.
"

Fetch the scale for each axis from sysfs for polling accelerometers.