main: Make it possible to instantiate multiple drivers
Split off "drivers" (class) and "devices" (instance) so that it's
possible to instantiate multiple devices of the same type, for example,
accelerometers in the base and the lid of laptops, or light sensors
in a removable keyboard as well as the display.

Note that we do not make use of that functionality just yet.