tests: Fix test_input_accel() leaking a daemon
test_input_accel() uses umockdev-run, and killing umockdev-run doesn't
kill the child iio-sensor-proxy daemon we make it spawn, making any test
running after this one fail. As the tests are run alphabetically, we
were lucky that this test was the last one, but we can't rely on that.

Instead, use psutil to list all the children processes and kill those
before killing umockdev-run.